backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# PDF Processing
pdfplumber==0.11.9
pytesseract==0.3.10
Pillow==12.1.0

# HTTP Client
requests==2.31.0

# Data Export
pandas==2.3.3
openpyxl==3.1.2

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
